Refactor usage of env vars
Using system-level environment variables won't work in the production build, so remove the usages of them.
Env vars through process.env only work in a Node environment like dev, but not when in the prod build which is just static content.
The env vars specify paths, and should still work if they're set to relative paths.

Fix HS-320