Disable triggers of individual workflows in favour of just using develop
This triggers only the develop workflow for PRs instead of individual, path-filtered workflows. Ideally we would use the path-filtered workflows and an e2e workflow that simply makes use of the built images, but we don't currently have a way for it to know what images it would need to use.

Fix HS-817