Create a helper include for deployment-specific env variables
The boilerplate include is put at the bottom of the env list so that it
is out of the way of more interesting deployment-specific stuff. Also,
this allows environment variables to be overridden in the values file
because it is at the end of the env list and the last occurence wins:
https://renehernandez.io/notes/env-vars-precedence-k8s-pod/

This also uses only the tag on the Docker image name, if it exists,
for the service.version resource attribute.