Tilt: fix adding test location to database
Follow the pattern in tests/minion-cert.sh for adding a location
(it it doesn't already exist) in the initContainer for the minion
that sets up certificates.

Added error checking so the init container won't successfully
finish until the location exists and the certificate is created.
In my experience, it takes 2-3 passes until everything is up, but
it does eventually complete.