minion: Add startupProbe that looks for a taskDefinition
Also now that we wait for the Minion service to startup, we need
to increase the timeout in install-local.sh when installing via
Helm with --wait to give time for downstream dependencies to come
up and for the Minion to come up.  When OpenNMS.Minion.addTestLocation
is true, Keycloak and Inventory are dependencies for the initContainer
and they each take a few minutes to come up. They don't currently
have any probes, so right now, they appear ready immediately, but it
take minutes for them to function. With the new Minion startupProbe,
we need to increase the timeout to allow for these dependencies.

Capture timing information when we run Helm so we can see how close
we are getting to the timeout.